+++ hs-overlay.nix
@@ -1,0 +1,3 @@
+self: super: {
+  pipes-zlib = super.callPackage ./pkg.nix {};
+}