+++ changelog.md
@@ -1,3 +1,8 @@
+# Version 0.4.4.2
+
+* Remove upper bound from all dependencies except `base`.
+
+
 # Version 0.4.4.1
 
 * Bump upper bound dependency on `pipes`.